decomp: add several libsnd functions (#44)

+++ decomp/src/libspu/s_m_util.c
@@ -7,16 +7,17 @@ s32 _SpuIsInAllocateArea(u32 arg0) {
         return 0;
     }
     for (i = 0;; i++) {
-        if (_spu_memList[i].addr & 0x80000000) {
+        if ((u_long)_spu_memList[i].addr & 0x80000000) {
             continue;
         }
-        if (_spu_memList[i].addr & 0x40000000) {
+        if ((u_long)_spu_memList[i].addr & 0x40000000) {
             break;
         }
-        if (arg0 <= (_spu_memList[i].addr & 0x0FFFFFFF)) {
+        if (arg0 <= ((u_long)_spu_memList[i].addr & 0x0FFFFFFF)) {
             return 1;
         }
-        if (arg0 < (_spu_memList[i].addr & 0x0FFFFFFF) + _spu_memList[i].size) {
+        if (arg0 < ((u_long)_spu_memList[i].addr & 0x0FFFFFFF) +
+                       _spu_memList[i].size) {
             return 1;
         }
     }
@@ -31,16 +32,17 @@ s32 _SpuIsInAllocateArea_(u32 arg0) {
         return 0;
     }
     for (i = 0; 1; i++) {
-        if (_spu_memList[i].addr & 0x80000000) {
+        if ((u_long)_spu_memList[i].addr & 0x80000000) {
             continue;
         }
-        if (_spu_memList[i].addr & 0x40000000) {
+        if ((u_long)_spu_memList[i].addr & 0x40000000) {
             break;
         }
-        if (arg0 <= (_spu_memList[i].addr & 0x0FFFFFFF)) {
+        if (arg0 <= ((u_long)_spu_memList[i].addr & 0x0FFFFFFF)) {
             return 1;
         }
-        if (arg0 < (_spu_memList[i].addr & 0x0FFFFFFF) + _spu_memList[i].size) {
+        if (arg0 < ((u_long)_spu_memList[i].addr & 0x0FFFFFFF) +
+                       _spu_memList[i].size) {
             return 1;
         }
     }
